Move files from visual to common
* DebugApolloError
* ErrorOverlay
* LoadingError
* LoadingOverlay
* Table
