+++ frontend/src/components/common/DebugApolloError.js
