+++ frontend/src/components/common/index.js
@@ -1,3 +1,7 @@
 // TODO: add more
 
 export {default as AdaValue} from './AdaValue'
+export {default as DebugApolloError} from './DebugApolloError'
+export {default as ErrorOverlay} from './ErrorOverlay'
+export {default as LoadingError} from './LoadingError'
+export {default as LoadingOverlay} from './LoadingOverlay'
