+++ frontend/src/components/visual/index.js
@@ -1,9 +1,6 @@
 // @flow
 export {default as LoadingInProgress} from './LoadingInProgress'
-export {default as LoadingOverlay} from './LoadingOverlay'
 export {default as Overlay} from './Overlay'
-export {default as ErrorOverlay} from './ErrorOverlay'
-export {default as DebugApolloError} from './DebugApolloError'
 export {default as SimpleLayout} from './SimpleLayout'
 export {default as SummaryCard} from './SummaryCard'
 export {default as Button} from './Button'
@@ -26,7 +23,6 @@ export {default as Alert} from './Alert'
 export {default as Searchbar} from './Searchbar'
 export {default as ToggleButton} from './ToggleButton'
 export {default as LoadingDots} from './LoadingDots'
-export {default as LoadingError} from './LoadingError'
 export {Tabs, Tab} from './Tabs'
 export {LiteTabs, LiteTab} from './LiteTabs'
 export {default as Divider} from './Divider'
