+++ frontend/src/screens/Blockchain/Address/PagedTransactions.js
@@ -9,7 +9,6 @@ import BigNumber from 'bignumber.js'
 
 import {
   LoadingInProgress,
-  LoadingError,
   EntityCardContent,
   SummaryCard,
   Link,
@@ -17,7 +16,7 @@ import {
   LiteTab,
   Card,
 } from '@/components/visual'
-import {AdaValue} from '@/components/common'
+import {AdaValue, LoadingError} from '@/components/common'
 import {getDefaultSpacing} from '@/components/visual/ContentSpacing'
 import {useI18n} from '@/i18n/helpers'
 import {routeTo} from '@/helpers/routes'
