+++ frontend/src/screens/Blockchain/Block/Slot.js
@@ -9,7 +9,8 @@ import {useQuery} from 'react-apollo-hooks'
 import gql from 'graphql-tag'
 
 import NotFound from '../NotFound'
-import {SimpleLayout, SummaryCard, LoadingDots, LoadingError} from '@/components/visual'
+import {SimpleLayout, SummaryCard, LoadingDots} from '@/components/visual'
+import {LoadingError} from '@/components/common'
 import {useI18n} from '@/i18n/helpers'
 import {isInteger} from '@/helpers/validators'
 import {routeTo} from '@/helpers/routes'
