+++ frontend/src/screens/Blockchain/Block/index.js
@@ -14,15 +14,14 @@ import {
   SimpleLayout,
   LoadingDots,
   LoadingInProgress,
-  LoadingError,
   Link,
 } from '@/components/visual'
+import {AdaValue, LoadingError} from '@/components/common'
 
 import blockIcon from '@/static/assets/icons/metrics-blocks.svg'
 import {useScrollFromBottom} from '@/components/hooks/useScrollFromBottom'
 import {useAnalytics} from '@/helpers/googleAnalytics'
 
-import {AdaValue} from '@/components/common'
 import TransactionCard from '@/components/common/TransactionCard'
 
 import {useI18n} from '@/i18n/helpers'
