+++ frontend/src/screens/Blockchain/Epoch/StakingPools.js
@@ -6,9 +6,9 @@ import {compose} from 'redux'
 import idx from 'idx'
 import {withI18n} from '@/i18n/helpers'
 import {routeTo} from '@/helpers/routes'
-import {LoadingInProgress, DebugApolloError, Link} from '@/components/visual'
-import {AdaValue} from '@/components/common'
-import Table, {ROW_TYPE} from '@/components/visual/Table'
+import {LoadingInProgress, Link} from '@/components/visual'
+import {AdaValue, DebugApolloError} from '@/components/common'
+import Table, {ROW_TYPE} from '@/components/common/Table'
 
 const GET_STAKE_POOLS_IN_EPOCH = gql`
   query($epochNumber: Int!) {
