+++ frontend/src/screens/Blockchain/Epoch/index.js
@@ -12,14 +12,12 @@ import {
   SimpleLayout,
   EntityCardContent,
   EntityCardShell,
-  LoadingError,
   Tab,
   Tabs,
   Overlay,
-  LoadingOverlay,
   Chip,
 } from '@/components/visual'
-import {AdaValue} from '@/components/common'
+import {AdaValue, LoadingError, LoadingOverlay} from '@/components/common'
 import {useI18n} from '@/i18n/helpers'
 import EpochNumberIcon from '@/static/assets/icons/epoch-number.svg'
 import EpochIcon from '@/static/assets/icons/metrics-epoch.svg'
