+++ frontend/src/screens/Blockchain/PagedBlocks/BlocksTable.js
@@ -4,8 +4,8 @@ import {defineMessages} from 'react-intl'
 import {Typography} from '@material-ui/core'
 import {makeStyles} from '@material-ui/styles'
 
-import Table, {ROW_TYPE} from '@/components/visual/Table'
 import {Link, Tooltip} from '@/components/visual'
+import Table, {ROW_TYPE} from '@/components/common/Table'
 import {AdaValue} from '@/components/common'
 import {useI18n} from '@/i18n/helpers'
 import {routeTo} from '@/helpers/routes'
