+++ frontend/src/screens/Blockchain/StakePool/index.js
@@ -6,13 +6,8 @@ import gql from 'graphql-tag'
 import {useQuery} from 'react-apollo-hooks'
 import useReactRouter from 'use-react-router'
 
-import {
-  EntityIdCard,
-  SummaryCard,
-  SimpleLayout,
-  LoadingInProgress,
-  LoadingError,
-} from '@/components/visual'
+import {EntityIdCard, SummaryCard, SimpleLayout, LoadingInProgress} from '@/components/visual'
+import {LoadingError} from '@/components/common'
 
 import {useScrollFromBottom} from '@/components/hooks/useScrollFromBottom'
 import {useI18n} from '@/i18n/helpers'
