+++ frontend/src/screens/Blockchain/StakingKey/User/index.js
@@ -4,15 +4,8 @@ import {Typography} from '@material-ui/core'
 import useReactRouter from 'use-react-router'
 import {makeStyles} from '@material-ui/styles'
 
-import {
-  SummaryCard,
-  SimpleLayout,
-  EntityIdCard,
-  Link,
-  LoadingError,
-  LoadingInProgress,
-} from '@/components/visual'
-import {AdaValue} from '@/components/common'
+import {SummaryCard, SimpleLayout, EntityIdCard, Link, LoadingInProgress} from '@/components/visual'
+import {AdaValue, LoadingError} from '@/components/common'
 import {routeTo} from '@/helpers/routes'
 import {useI18n} from '@/i18n/helpers'
 import AdaIcon from '@/static/assets/icons/transaction-id.svg'
