+++ frontend/src/screens/Blockchain/Transaction/index.js
@@ -13,13 +13,11 @@ import {
   SummaryCard,
   SimpleLayout,
   LoadingInProgress,
-  LoadingError,
   EntityIdCard,
   Link,
   Overlay,
-  LoadingOverlay,
 } from '@/components/visual'
-import {AdaValue} from '@/components/common'
+import {AdaValue, LoadingError, LoadingOverlay} from '@/components/common'
 import AddressesBreakdown from '@/components/common/AddressesBreakdown'
 import AssuranceChip from '@/components/common/AssuranceChip'
 import AdaIcon from '@/static/assets/icons/transaction-id.svg'
