+++ frontend/src/screens/Home/Charts/index.js
@@ -13,15 +13,8 @@ import {makeStyles} from '@material-ui/styles'
 import {useI18n} from '@/i18n/helpers'
 import {ObjectValues} from '@/helpers/flow'
 import {useAnalytics} from '@/helpers/googleAnalytics'
-import {
-  SimpleLayout,
-  LiteTabs,
-  LiteTab,
-  Card,
-  LoadingInProgress,
-  LoadingError,
-  Alert,
-} from '@/components/visual'
+import {SimpleLayout, LiteTabs, LiteTab, Card, LoadingInProgress, Alert} from '@/components/visual'
+import {LoadingError} from '@/components/common'
 import useTabState from '@/components/hooks/useTabState'
 import {useCurrentBreakpoint} from '@/components/hooks/useBreakpoints'
 import {TabsProvider as Tabs, TabItem as Tab, useTabContext} from '@/components/context/TabContext'
