+++ frontend/src/screens/Staking/SideMenu/PoolsToCompare.js
@@ -12,7 +12,8 @@ import {fade} from '@material-ui/core/styles/colorManipulator'
 import {IconButton, Grid, Typography, Avatar, withStyles} from '@material-ui/core'
 import {Share, CallMade, CallReceived, Clear} from '@material-ui/icons'
 
-import {LoadingDots, DebugApolloError, VisualHash, Chip} from '@/components/visual'
+import {LoadingDots, VisualHash, Chip} from '@/components/visual'
+import {DebugApolloError} from '@/components/common'
 import CopyToClipboard from '@/components/common/CopyToClipboard'
 import assert from 'assert'
 import {withI18n} from '@/i18n/helpers'
