+++ frontend/src/screens/Staking/StakeList/index.js
@@ -9,7 +9,8 @@ import {withHandlers, withProps} from 'recompose'
 import {defineMessages} from 'react-intl'
 
 import {useI18n} from '@/i18n/helpers'
-import {Button, Overlay, LoadingInProgress, LoadingOverlay, LoadingError} from '@/components/visual'
+import {Button, Overlay, LoadingInProgress} from '@/components/visual'
+import {LoadingError, LoadingOverlay} from '@/components/common'
 import StakePool from './StakePool'
 import SearchAndFilterBar from './SearchAndFilterBar'
 import SortByBar from './SortByBar'
